requirements.txt
# Python 3.12+ required

# Core dependencies
requests==2.31.0          # HTTP client for API calls
beautifulsoup4==4.12.3    # HTML parsing for article content
markdownify==1.1.0        # Convert HTML to Markdown
openai==1.76.2             # OpenAI API client for vector stores
python-dotenv==1.0.1      # Environment variable management
tiktoken==0.9.0            # OpenAI tokenizer for accurate chunk estimation
